+++ Cvdm.ErrorHandling.Tests/AsyncResultBuilderTests.fs
@@ -1061,3 +1061,27 @@ let ``use! ignores null async-wrapped disposable`` () =
       } |> Async.RunSynchronously
     test <@ result = Ok () @>
   }
+
+[<Fact>]
+let ``use! handles non-nullable async-wrapped disposable`` () =
+  Property.check <| property {
+    raises<CustomDisposedException>
+      <@      
+        asyncResult {
+          use! _d = async { return (new CustomDisposable()) }
+          do! Ok ()
+        } |> Async.RunSynchronously
+      @>
+  }
+
+[<Fact>]
+let ``use handles non-nullable async-wrapped disposable`` () =
+  Property.check <| property {
+    raises<CustomDisposedException>
+      <@       
+        asyncResult {
+          use _d = new CustomDisposable()
+          do! Ok ()
+        } |> Async.RunSynchronously
+      @>
+  }